+++ concordium-sdk/src/main/java/com/concordium/sdk/ClientV2MapperExtensions.java
@@ -1,8 +1,11 @@
 package com.concordium.sdk;
 
+import com.concordium.grpc.v2.AccountAddress;
 import com.concordium.grpc.v2.AccountInfo;
 import com.concordium.grpc.v2.Commitment;
 import com.concordium.grpc.v2.CredentialPublicKeys;
+import com.concordium.grpc.v2.CredentialRegistrationId;
+import com.concordium.grpc.v2.EncryptedAmount;
 import com.concordium.grpc.v2.Policy;
 import com.concordium.grpc.v2.ReleaseSchedule;
 import com.concordium.grpc.v2.*;
@@ -20,10 +23,7 @@ import com.concordium.sdk.responses.blocksummary.updates.queues.AnonymityRevoker
 import com.concordium.sdk.responses.blocksummary.updates.queues.Description;
 import com.concordium.sdk.responses.transactionstatus.DelegationTarget;
 import com.concordium.sdk.responses.transactionstatus.OpenStatus;
-import com.concordium.sdk.transactions.CCDAmount;
-import com.concordium.sdk.transactions.EncryptedAmountIndex;
-import com.concordium.sdk.transactions.Hash;
-import com.concordium.sdk.transactions.Index;
+import com.concordium.sdk.transactions.*;
 import com.concordium.sdk.types.Nonce;
 import com.concordium.sdk.types.Timestamp;
 import com.concordium.sdk.types.UInt64;
@@ -423,6 +423,11 @@ interface ClientV2MapperExtensions {
                         to(schedule.getSchedulesList(), ClientV2MapperExtensions::to)))
                 .build();
     }
+
+    static AccountNonce to(NextAccountSequenceNumber nextAccountSequenceNumber) {
+        return AccountNonce.from(to(nextAccountSequenceNumber.getSequenceNumber()));
+    }
+
 
     static <T1, T2> List<T2> to(List<T1> sourceList, Function<T1, T2> map) {
         return sourceList.stream().map(map::apply).collect(Collectors.toList());